+++ Sora/MediaChannel.swift
@@ -279,6 +279,16 @@ public final class MediaChannel {
     connectionStartTime = nil
     connectionTask.peerChannel = peerChannel
 
+    // NativePeerChannelFactory がシングルトンのため
+    // 前回切断時に音声ハードミュート状態の場合は状態が残ってしまうので
+    // 音声ハードミュートのリセットを行います
+    if configuration.audioEnabled {
+      Logger.debug(
+        type: .mediaChannel,
+        message: "reset audio hard mute before connecting")
+      _ = NativePeerChannelFactory.default.audioDeviceModuleWrapper.resetHardMuteIfNeeded()
+    }
+
     signalingChannel.internalHandlers.onDisconnect = { [weak self] error, reason in
       guard let weakSelf = self else {
         return
@@ -520,6 +530,27 @@ public final class MediaChannel {
 
     return result
       ? nil : SoraError.messagingError(reason: "failed to send message: label => \(label)")
+  }
+
+  /// MediaChannel の接続中にマイクをハードミュート有効化/無効化します
+  /// - Parameter mute: `true` で有効化、`false` で無効化
+  /// - Returns: 成功した場合は `true`、接続状態などの理由で処理しなかった場合は `false`
+  public func setAudioHardMute(_ mute: Bool) -> Bool {
+    guard state == .connected else {
+      Logger.debug(
+        type: .mediaChannel,
+        message: "setAudioHardMute failed, cause MediaChannel is not connected (state: \(state))")
+      return false
+    }
+
+    guard configuration.audioEnabled else {
+      Logger.debug(
+        type: .mediaChannel,
+        message: "setAudioHardMute skipped because audioEnabled is false")
+      return false
+    }
+
+    return NativePeerChannelFactory.default.audioDeviceModuleWrapper.setAudioHardMute(mute)
   }
 }